Use traits for API operations methods
