+++ lib/ApiOperations/Retrieve.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Stripe\ApiOperations;
+
+/**
+ * Trait for retrievable resources. Adds a `retrieve()` static method to the
+ * class.
+ *
+ * This trait should only be applied to classes that derive from ApiResource.
+ */
+trait Retrieve
+{
+    /**
+     * @param array|string $id The ID of the API resource to retrieve,
+     *     or an options array containing an `id` key.
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function retrieve($id, $opts = null)
+    {
+        return self::_retrieve($id, $opts);
+    }
+}
