+++ lib/ApiOperations/Update.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Stripe\ApiOperations;
+
+/**
+ * Trait for updatable resources. Adds an `update()` static method and a
+ * `save()` method to the class.
+ *
+ * This trait should only be applied to classes that derive from ApiResource.
+ */
+trait Update
+{
+    /**
+     * @param string $id The ID of the resource to update.
+     * @param array|null $params
+     * @param array|string|null $options
+     *
+     * @return ApiResource The updated resource.
+     */
+    public static function update($id, $params = null, $options = null)
+    {
+        return self::_update($id, $params, $options);
+    }
+
+    /**
+     * @param array|string|null $options
+     *
+     * @return ApiResource The saved resource.
+     */
+    public function save($options = null)
+    {
+        return $this->_save($options);
+    }
+}
