+++ lib/BalanceTransaction.php
@@ -24,6 +24,9 @@ namespace Stripe;
  */
 class BalanceTransaction extends ApiResource
 {
+    use ApiOperations\All;
+    use ApiOperations\Retrieve;
+
     /**
      * @return string The class URL for this resource. It needs to be special
      *    cased because it doesn't fit into the standard resource pattern.
@@ -31,28 +34,5 @@ class BalanceTransaction extends ApiResource
     public static function classUrl()
     {
         return "/v1/balance/history";
-    }
-
-    /**
-     * @param array|string $id The ID of the balance transaction to retrieve,
-     *     or an options array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return BalanceTransaction
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of BalanceTransactions
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
     }
 }
