+++ lib/CountrySpec.php
@@ -9,6 +9,9 @@ namespace Stripe;
  */
 class CountrySpec extends ApiResource
 {
+    use ApiOperations\All;
+    use ApiOperations\Retrieve;
+
     /**
      * This is a special case because the country specs endpoint has an
      *    underscore in it. The parent `className` function strips underscores.
@@ -18,29 +21,5 @@ class CountrySpec extends ApiResource
     public static function className()
     {
         return 'country_spec';
-    }
-
-    /**
-     * @param array|string $country The ISO country code of the country we
-     *     retrieve the country specfication for, or an options array
-     *     containing an `id` containing that code.
-     * @param array|string|null $opts
-     *
-     * @return CountrySpec
-     */
-    public static function retrieve($country, $opts = null)
-    {
-        return self::_retrieve($country, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of CountrySpecs
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
     }
 }
