+++ lib/ExchangeRate.php
@@ -9,6 +9,9 @@ namespace Stripe;
  */
 class ExchangeRate extends ApiResource
 {
+    use ApiOperations\All;
+    use ApiOperations\Retrieve;
+
     /**
      * This is a special case because the exchange rates endpoint has an
      *    underscore in it. The parent `className` function strips underscores.
@@ -18,27 +21,5 @@ class ExchangeRate extends ApiResource
     public static function className()
     {
         return 'exchange_rate';
-    }
-
-    /**
-     * @param array|string $currency
-     * @param array|string|null $opts
-     *
-     * @return ExchangeRate
-     */
-    public static function retrieve($currency, $opts = null)
-    {
-        return self::_retrieve($currency, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return ExchangeRate
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
     }
 }
