+++ lib/Subscription.php
@@ -9,6 +9,11 @@ namespace Stripe;
  */
 class Subscription extends ApiResource
 {
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
+
     /**
      * These constants are possible representations of the status field.
      *
@@ -21,52 +26,6 @@ class Subscription extends ApiResource
     const STATUS_UNPAID = 'unpaid';
 
     /**
-     * @param array|string $id The ID of the subscription to retrieve, or an
-     *     options array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Subscription
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Subscriptions
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Subscription The created subscription.
-     */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
-
-    /**
-     * @param string $id The ID of the subscription to retrieve.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Subscription The updated subscription.
-     */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
      * @param array|null $params
      *
      * @return Subscription The deleted subscription.
@@ -74,16 +33,6 @@ class Subscription extends ApiResource
     public function cancel($params = null, $opts = null)
     {
         return $this->_delete($params, $opts);
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return Subscription The saved subscription.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
     }
 
     /**
